Draft for POC Design that includes completed sections
This includes completion of the user stories in ethnography, themes, categories and prerequisites of user stories and acceptance criteria. 

Additional editing may be needed after feedback is given for prior elaboration submissions. You will need to check whether you must be discussing a specific pre-existing technology or an abstract one of your own. This is most likely crucial to progress.